requirements.txt
azure-core==1.36.0 ; python_version >= "3.12" and python_version < "4.0"
azure-identity-broker==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
azure-identity==1.25.1 ; python_version >= "3.12" and python_version < "4.0"
azure-monitor-query==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.12" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
contourpy==1.3.3 ; python_version >= "3.12" and python_version < "4.0"
cryptography==46.0.3 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
fonttools==4.60.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.12" and python_version < "4.0"
jinja2>=3.1.0 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.9 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.10.7 ; python_version >= "3.12" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
msal==1.34.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.3.5 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.3.3 ; python_version >= "3.12" and python_version < "4.0"
pillow==12.0.0 ; python_version >= "3.12" and python_version < "4.0"
plotly==5.24.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pymsalruntime==0.18.1 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or platform_system == "Darwin" or platform_system == "Linux")
pyparsing==3.2.5 ; python_version >= "3.12" and python_version < "4.0"
pyyaml>=6.0.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
tenacity==9.1.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "4.0"
